initialization of defined variables
BZ: 63404
[CLEAN]

Because in functions of hardware.c, the variables are
not initialized when they are defined, valgrind indicates
that these variables are used as uninitialised! So,
the purpose of this patch is initialization of varaibles.
In the same way, the patch fixes the compilation warning
(unused variable).

Change-Id: I7c1db51c155f064dc57ac417a12cc0030cc278af
Orig-Change-Id: I3ed3608ed1bf468e275e01c560a2d3f1319cdedd
Category: aosp improvement
Domain: AudioComms-HwModules-Audio
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>