[PORT FROM MAIN] initialization of defined variables
BZ: 63404

Because in functions of hardware.c, the variables are
not initialized when they are defined, valgrind indicates
that these variables are used as uninitialised! So,
the purpose of this patch is initialization of varaibles.
In the same way, the patch fixes the compilation warning
(unused variable).

Feature: None
Category: Improvement
Purpose: AOSP+
Origin: Internal
Upstream-candidate: yes
Change-Id: I3ed3608ed1bf468e275e01c560a2d3f1319cdedd
Orig-Change-Id: Icfadf6e2fe2206d05a9c48682b10119e5e4d4fdb
Signed-off-by: Andrei V. FOMITCHEV <andreix.fomitchev@intel.com>