PowerHAL[REVERTME]: Disabling timer rate tuning and vsync boost on touch
BZ: 79405

This patch temporarily disables the timer rate tuning and vsync boost.
On MR1 release, the Power HAL functionality has changed and so the
module needs a redesign. But we need to temporarily disable the touch
boost features already implemented for MR0 as the governor timer rate
is locked to 20ms in MR1 with the same Power HAL as in MR0.
The impact on keeping the timer rate at 20 ms is 1)power regressions
2) performance improvement.
This patch would actually help to keep the timer rate at the optimal
value of 100ms tuning the system to optimal power.

Change-Id: Ia88e560d079ad04391c9db32308ee0163bda83c9
Signed-off-by: Mahesh Kumar <mahesh.kumar.p@intel.com>
Signed-off-by: Ashish K <ashish.k@intel.com>
Reviewed-on: http://android.intel.com:8080/86200
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Tested-by: Cuesta, FernandX <fernandx.cuesta@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>