PowerHAL: Touch and Vsync boost implementation on MR1 build
BZ: 82040

The Power HAL implementation was brokenon MR1 build due to google architectural changes.
This patch does the reimplementation of Power HAL.

On MR1 ,the vsync hint and touch hints are triggered from two different process unlike MR0.
To account for this change in Android release, the Power HAL touch and vsync boost is
reimplemented.

Change-Id: Ieee083684946cd89b0bbcd0b4b7ffb4535961b7d
Signed-off-by: Kumar Mahesh <mahesh.kumar.p@intel.com>
Reviewed-on: http://android.intel.com:8080/88001
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Mansoor, Illyas <illyas.mansoor@intel.com>
Tested-by: Bourahmani, KarimX <karimx.bourahmani@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>