[AUDIO] Add a setParameter from AudioSystem to the Policy
BZ: 163953

Some services use setParameter of AudioSystem to inform
of platform changes. This platform changes cannot be sent to the
policy. FM Rx playback requires for example the policy to be aware of the
requested output device and must inform the policy.

This patch adds a setParameter link from AudioSystem to the policy
to handle generic setParameter calls in the policy if needed.

Change-Id: Ifb6ed2bbae0c38feea5589e4a3e45a1a002d797c
Category: device enablement
Domain: AudioComms-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>