[PORT FROM MAIN] [REMOTE BGM][DO NOT UPSTREAM] Local audio routes to remote sink when playing video only stream on remote sink
BZ: 101834

The audio policy routing to the remote sink gets initialized
only on first successful playback of an audio stream. In case
of video only stream, the audio path is not hit. So this
information is transmitted to HAL and retrieved in APM for
routing the audio to appropriate sinks.

Change-Id: I6f806cbd2c4a2271914bbebd30cd005fbda317f5
Orig-Change-Id: I5a776a6913b22db251c17b5d840c419b2c37d42c
Signed-off-by: jerome <jerome.anand@intel.com>
Reviewed-on: http://android.intel.com:8080/104287
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Jayanti, Satya Charitardha <satya.charitardha.jayanti@intel.com>
Reviewed-by: Sikkandar D, Madar <madar.sikkandar.d@intel.com>
Reviewed-by: Krishna, Samaga <samaga.krishna@intel.com>
Tested-by: Joshi, SantoshX <santoshx.joshi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>