[PORT FROM MAIN]Introduce POWER_HINT_PAGELOADING into power HAL
BZ: 79423

Browser has started or stopped reqesting the pageloading, i.e.
POWER_HINT_PAGELOADING. If the browser has requested pageloading,
then CPU and GPU load is expected soon, and it may be appropriate
to raise speeds of CPU, memory bus, etc.

For the hint to take effect, it relies on the implementations:
1. PowerHAL should implement an interface to let browser invoke,
 the hint. When received the hint, PowerHAL should response it with
some corresponsing system behavior.
2. Browser webkit should have the implementation, call the hint
interface defined in PowerHAL when pageloading is stared.

With all the implementations, this hint can help improve
pageloading speed. BrowsingBench score is improved from ~2900 to
~3300 in stock browser. And power data shows no impact to Browser_I
power.

Change-Id: I3d375e6f54235a6a7f44407c43042abaa3c92eaa
Signed-off-by: cwang99 <chuan.a.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/122564
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>