gralloc: framebuffer: support BGR framebuffer
The GL subsystem, PixelFlinger and default gralloc all assume
the framebuffer is in RGB{_,X,A}. When the framebuffer itself
is in BGRA format, gralloc cannot simply do memcpy() to copy
the rendering buffer to the framebuffer. Therefore, channel
swapping needs to be done when posting the rendering buffer.
The swapping is done for i915 framebuffer with modeset, and
for EFI framebuffer if forced. The efifb driver always reports
RGB so BGR has to be forcibly enabled.

Note that, instead of having PixelFlinger swapping channels
during screen composition, the swapping is done when a rendering
buffer is being copied to the actual framebuffer. This should
cause less of performance degradation since the conversion is not
done for individual textures. This also avoids overhauling
both libagl and PixelFlinger to make them BGR-compatible.

Category: AOSP Enhancement
Domain: Gralloc, Software Rendering
Origin: Internal
Upstream-Candidate: Yes

For: ADEV-310
Change-Id: I98735074cde117b02a84e9b8961a303d1da9ebcd
Signed-off-by: Daniel Leung <daniel.leung@intel.com>