Add vibrator's minimum timeout to vibrator_module structure
BZ: 142498

Depending on the hardware type, vibrations under certain threshold cannot be
felt. This is mainly due to the vibrator slow response or the lag time.

To compensate this, the HAL will now fetch the minimum timeout allowed for a
short vibration from the parameter framework and forward it to upper layers.
This patch adds the minimum timeout parameter to vibrator_module structure.

Change-Id: Iac9e0d0ef2a2407df0470637c1d9f12c9d37b85e
Category: aosp improvement
Domain: AudioComms-HwModules-Vibrator
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Piotr Diop <piotrx.diop@intel.com>