[EM/BCU_HAL]: Fix for the Klockwork issue #15367
BZ: 174688

Issue Description:
The allocated buffer size is 4, but while copying data using sprintf function
the size is unknown bytes, because of this buffer gets overflows.
Solution:
The sprintf function, which simply assumes that the output buffer is large
enough to hold the resulting string, was replaced with the snprintf function,
which writes a maximum number of bytes to the buffer. This (snprintf function)
will restrict the data while copying within the limit of max buffer size. So the
buffer overflow will not occur.

Change-Id: I42d5a0a25a3b9039d0239e4c5d891c388a86258b
Category: device enablement
Domain: BatteryManagement-Common
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Albin B <albin.bala.krishnan@intel.com>