gralloc: force reported xres to be same as line_length
For some framebuffers, its advertised line_length does not correspond
to the width of the image. For example, 1366x768 display may have
width of 1368, but line_length is 5504 (1376 * 4). However, the 1368
value is reported to the upper layer. Copying the buffer to fb will
cause skewed image. This forces the reported xres to correspond to
line_length.

Change-Id: I7e9982d53256979ef52f32afc6f1d70eefd95653
Signed-off-by: Daniel Leung <daniel.leung@intel.com>
Reviewed-on: https://android.intel.com/177472
Reviewed-by: Uyyala, Sridhar <sridhar.uyyala@intel.com>
Tested-by: Uyyala, Sridhar <sridhar.uyyala@intel.com>
Reviewed-by: Bergeron, Michael <michael.bergeron@intel.com>
Tested-by: Bergeron, Michael <michael.bergeron@intel.com>
Reviewed-by: Kumar-mayernik, Nisha <nisha.kumar-mayernik@intel.com>
Reviewed-by: Gumbel, Matthew K <matthew.k.gumbel@intel.com>